+++ src/llist.h
@@ -8,12 +8,12 @@ typedef struct llist llist;
 
 llist *LinkedList_Create(void);
 void LinkedList_Destroy(llist *self);
-void LinkedList_Prepend(llist *self, uint16_t item);
-void LinkedList_Append(llist *self, uint16_t item);
-void LinkedList_DeleteItem(llist *self, uint16_t item);
+void LinkedList_Prepend(llist *self, uint32_t item);
+void LinkedList_Append(llist *self, uint32_t item);
+void LinkedList_DeleteItem(llist *self, uint32_t item);
 void LinkedList_Show(llist *self);
 bool LinkedList_IsEmpty(llist *self);
 uint32_t LinkedList_GetSize(llist *self);
-uint16_t LinkedList_Shift(llist *self);
+uint32_t LinkedList_Shift(llist *self);
 
 #endif // _LLIST_H
